+++ Arduino-Code/WiFi_Windsensor/error_html.h
@@ -22,7 +22,7 @@ String Error()
  
  // Web page title
  content +=F( "<h2>");
- content += String(actconf.devname);
+ content += String(actconf.devname) + " " + String(actconf.windSensorType);
  content +=F( "</h2>");
  content += String(actconf.crights);
  content +=F( ", ");