+++ Arduino-Code/WiFi_Windsensor/firmware_html.h
@@ -39,7 +39,7 @@ String Firmware(int num, String vname[20], String value[20])
    
    // Web page title
    content +=F( "<h2>");
-   content += String(actconf.devname);
+   content += String(actconf.devname) + " " + String(actconf.windSensorType);
    content +=F( "</h2>");
    content += String(actconf.crights);
    content +=F( ", "); 
@@ -186,7 +186,7 @@ String Firmware(int num, String vname[20], String value[20])
    
    // Web page title
    content +=F( "<h2>");
-   content += String(actconf.devname);
+   content += String(actconf.devname) + " " + String(actconf.windSensorType);
    content +=F( "</h2>");
    content += String(actconf.crights);
    content +=F( ", ");