+++ Arduino-Code/WiFi_Windsensor/main_html.h
@@ -80,7 +80,7 @@ String Startpage(int num, String vname[20], String value[20])
  
  // Web page title
  content +=F( "<h2>");
- content += String(actconf.devname);
+ content += String(actconf.devname) + " " + String(actconf.windSensorType);
  content +=F( "</h2>");
  content += String(actconf.crights);
  content +=F( ", ");